requirements.txt
# Deep Learning
--find-links https://download.pytorch.org/whl/torch_stable.html
--extra-index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision & SfM
opencv-python>=4.8.0
kornia>=0.7.0

# Feature Matching (State-of-the-art)
lightglue @ git+https://github.com/cvg/LightGlue.git

# Numerical Computing
numpy>=1.24.0
scipy>=1.10.0

# Configuration & Logging
hydra-core>=1.3.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
imageio>=2.31.0
pillow>=10.0.0
